# Faster similar for SizedArray