--- a/src/util.jl
+++ b/src/util.jl
@@ -43,7 +43,7 @@ end
 
 Converts a size given by `Tuple{N, M, ...}` into a tuple `(N, M, ...)`.
 """
-@generated function size_to_tuple(::Type{S}) where S<:Tuple
+@pure function size_to_tuple(::Type{S}) where S<:Tuple
     return tuple(S.parameters...)
 end