This small change speeds up `similar` for `SizedArray`.

Before:
```julia
julia> using StaticArrays

julia> using BenchmarkTools

julia> a = SizedArray{Tuple{2,2}}([1.0 2.0; 3.0 4.0])
2×2 SizedArray{Tuple{2,2},Float64,2,2}:
 1.0  2.0
 3.0  4.0

julia> @benchmark similar($a)
BenchmarkTools.Trial: 
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     177.235 ns (0.00% GC)
  median time:      179.944 ns (0.00% GC)
  mean time:        206.715 ns (8.87% GC)
  maximum time:     90.985 μs (99.74% GC)
  --------------
  samples:          10000
  evals/sample:     714
```

After:

```julia

julia> @benchmark similar($a)
BenchmarkTools.Trial: 
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     40.421 ns (0.00% GC)
  median time:      45.254 ns (0.00% GC)
  mean time:        64.077 ns (24.18% GC)
  maximum time:     63.615 μs (99.86% GC)
  --------------
  samples:          10000
  evals/sample:     992
```

My `versioninfo`:
```
julia> versioninfo()
Julia Version 1.1.0
Commit 80516ca202 (2019-01-21 21:24 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: Intel(R) Core(TM) i7-4800MQ CPU @ 2.70GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, haswell)
```