@c42f 2019-07-02
I checked the other places where we splat `parameters...`. All the other locations are inside `@generated` or `@pure`.