@mateuszbaran 2019-07-02
OK, I've changed it to `@pure`.